moved handlers for buttons to HostDevButtons component, passed emitter, setter, and state to that component from room
